implement abort and exit for c64 & a800xl
