+++ mos-platform/a800xl/lib/abort.c
@@ -1,0 +1,7 @@
+#include <stdlib.h>
+#include <stdio.h>
+
+void abort(void) {
+   puts("Aborted");
+   exit(134);  // 128 + SIGABRT
+}
