+++ mos-platform/sim/include/stdlib.h
@@ -7,9 +7,6 @@
 extern "C" {
 #endif
 
-void abort(void);
-
-void exit(int status);
 unsigned long clock();
 void reset_clock();
