+++ mos-platform/sim/lib/stdlib.c
@@ -10,16 +10,6 @@ void abort(void) {
   __builtin_unreachable();
 }
 
-void exit(int status) {
-  // Writing to this IO register causes the simulator to exit with the given status.
-  (*(volatile char *)0xFFF8) = (char)status;
-
-  // Prevent the compiler from considering this entire basic block unreachable.
-  asm volatile ("");
-
-  __builtin_unreachable();
-}
-
 unsigned long clock() {
   // reading first byte latches whole 32-bit value
   // so it needs to be read byte by byte in correct order
